feat(inline-wp): add inline work package chip component
