+++ lib/components/InlineWorkPackage/index.ts
@@ -1,0 +1,5 @@
+export { inlineWorkPackageSpec } from "./spec";
+export { WpOptionsPopover } from "./popovers";
+export { registerInlineWpCallbacks, clearInlineWpCallbacks } from "./callbacks";
+export type { InlineWpSize } from "./types";
+export type { WpOptionsProps } from "./popovers";
