+++ lib/components/index.ts
@@ -1,5 +1,4 @@
 export * from "./OpenProjectWorkPackageBlock";
 export * from "./ShadowDomWrapper";
-export * from "./InlineWorkPackageSpec";
+export * from "./InlineWorkPackage"; 
 export * from "./InlineWorkPackageSlashMenu";
-export * from "./InlineWorkPackageClipboardExtension";
