+++ src/App.tsx
@@ -13,10 +13,8 @@ import {
   openProjectWorkPackageBlockSpec,
   inlineWorkPackageSpec,
   inlineWorkPackageSlashMenu,
-  InlineWorkPackageClipboardExtension,
-  setLastPasteHtml,
 } from "../lib";
-import { useEffect, useRef } from "react";
+import { useEffect } from "react";
 import "./fetchOverride";
 
 const schema = BlockNoteSchema.create().extend({
@@ -31,14 +29,7 @@ const schema = BlockNoteSchema.create().extend({
 type EditorType = typeof schema.BlockNoteEditor;
 
 export default function App() {
-  const editorContainerRef = useRef<HTMLDivElement>(null);
-
-  const editor = useCreateBlockNote({
-    schema,
-    _tiptapOptions: {
-      extensions: [InlineWorkPackageClipboardExtension],
-    },
-  });
+  const editor = useCreateBlockNote({ schema });
 
   useEffect(() => {
     initializeOpBlockNoteExtensions({
@@ -47,36 +38,19 @@ export default function App() {
     });
   }, []);
 
-  useEffect(() => {
-    const container = editorContainerRef.current;
-    if (!container) return;
-
-    const onPaste = (e: ClipboardEvent) => {
-      const html = e.clipboardData?.getData('text/html') ?? '';
-      if (html.includes('data-inline-wp')) {
-        setLastPasteHtml(html);
-      }
-    };
-
-    container.addEventListener('paste', onPaste, { capture: true });
-    return () => container.removeEventListener('paste', onPaste, { capture: true });
-  }, []);
-
   const getCustomSlashMenuItems = (editorInstance: EditorType) => [
     ...getDefaultReactSlashMenuItems(editorInstance),
     inlineWorkPackageSlashMenu(editorInstance as any),
   ];
 
   return (
-    <div ref={editorContainerRef}>
-      <BlockNoteView editor={editor} slashMenu={false}>
-        <SuggestionMenuController
-          triggerCharacter="/"
-          getItems={async (query: string) =>
-            filterSuggestionItems(getCustomSlashMenuItems(editor), query)
-          }
-        />
-      </BlockNoteView>
-    </div>
+    <BlockNoteView editor={editor} slashMenu={false}>
+      <SuggestionMenuController
+        triggerCharacter="/"
+        getItems={async (query: string) =>
+          filterSuggestionItems(getCustomSlashMenuItems(editor), query)
+        }
+      />
+    </BlockNoteView>
   );
 }
